Revert "Merge branch 'master' into update/photo-system"
This reverts commit 488718290a556768aa3e42cf4590bf76470f24e7, reversing
changes made to 19ad6dd692d51f81c3f49485b6ada20985147a64.

+++ app/Http/Controllers/NarrowcastingController.php
@@ -42,6 +42,7 @@ class NarrowcastingController extends Controller
     {
         if (! $request->file('image') && ! $request->has('youtube_id')) {
             Session::flash('flash_message', 'Every campaign needs either an image or a video!');
+
             return Redirect::back();
         }
 
@@ -63,14 +64,15 @@ class NarrowcastingController extends Controller
         if ($request->has('youtube_id') && strlen($youtube_id) > 0) {
             $video = Youtube::getVideoInfo($youtube_id);
 
-            /* @phpstan-ignore-next-line */
             if (! $video) {
                 Session::flash('flash_message', 'This is an invalid video ID!');
+
                 return Redirect::back();
             }
 
             if (! $video->status->embeddable) {
                 Session::flash('flash_message', 'This video is not embeddable and therefore cannot be used on the site!');
+
                 return Redirect::back();
             }
 
@@ -91,9 +93,13 @@ class NarrowcastingController extends Controller
      */
     public function edit($id)
     {
-        $narrowcasting = NarrowcastingItem::findOrFail($id);
+        $narrowcasting = NarrowcastingItem::find($id);
 
-        return view('narrowcasting.edit', ['item' => $narrowcasting]);
+        if ($narrowcasting) {
+            return view('narrowcasting.edit', ['item' => $narrowcasting]);
+        } else {
+            abort(404);
+        }
     }
 
     /**
@@ -104,7 +110,11 @@ class NarrowcastingController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $narrowcasting = NarrowcastingItem::findOrFail($id);
+        $narrowcasting = NarrowcastingItem::find($id);
+
+        if (! $narrowcasting) {
+            abort(404);
+        }
 
         $narrowcasting->name = $request->name;
         $narrowcasting->campaign_start = strtotime($request->campaign_start);
@@ -123,9 +133,9 @@ class NarrowcastingController extends Controller
         if ($request->has('youtube_id') && strlen($youtube_id) > 0) {
             $video = Youtube::getVideoInfo($youtube_id);
 
-            /* @phpstan-ignore-next-line */
             if (! $video) {
                 Session::flash('flash_message', 'This is an invalid video ID!');
+
                 return Redirect::back();
             }
 
@@ -149,7 +159,11 @@ class NarrowcastingController extends Controller
      */
     public function destroy($id)
     {
-        $narrowcasting = NarrowcastingItem::findOrFail($id);
+        $narrowcasting = NarrowcastingItem::find($id);
+
+        if (! $narrowcasting) {
+            abort(404);
+        }
 
         Session::flash('flash_message', "Your campaign '".$narrowcasting->name."' has been deleted.");
         $narrowcasting->delete();
@@ -168,6 +182,7 @@ class NarrowcastingController extends Controller
         }
 
         Session::flash('flash_message', 'All finished campaigns have been deleted.');
+
         return Redirect::route('narrowcasting::list');
     }
